requirements.txt
# Core dependencies for IMDB Dataset Explorer
numpy==1.24.3
scipy==1.10.1
pandas==1.5.3
python-dateutil==2.8.2
# Fix Flask and Werkzeug versions for compatibility
flask==2.0.1
werkzeug==2.0.3

